Rename debugger module to debug
